third_party/blink/web_tests/webaudio/BiquadFilter/biquad-829349.html

<!doctype html>
<html>
  <head>
    <title>
      Biquad Tailtime Computation (crbug.com/829349)
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
  </head>
  <body>
    <script>
      // Sample rate used in testing crbug.com/829349.  Use the same rate to
      // make sure the coefficients have the same values.  (The actual
      // coefficients depend on the sample rate, so keep it the same for
      // consistency with the repro case.)
      const sampleRate = 44100;

      promise_test(async () => {
        // Short length
        const context =
            new OfflineAudioContext({length: 1024, sampleRate: sampleRate});
        const src = new OscillatorNode(context);

        // This particular filter is one of the filters used in the repro case
        // from crbug.com/829349.  In a debug build, this causes a DCHECK crash
        // because a value as NaN.  Just use this filter and verify that it
        // works and doesn't crash in a debug build.
        const f = new BiquadFilterNode(
            context,
            {type: 'peaking', frequency: 15000, Q: 0.2, gain: -5, detune: 0});

        src.connect(f).connect(context.destination);

        src.start();
        const audioBuffer = await context.startRendering();

        // Just verify that the output is not all zeroes.  The actual test
        // is that we don't have a DCHECK failure in a debug build.
        const data = audioBuffer.getChannelData(0);
        assert_not_constant_value(
            data, 0,
            'Filter output should not be all zeroes');
      }, 'Peaking filter tailtime');
    </script>
  </body>
</html>
